Keep refactoring views
