+++ BalanceApi/Model/Views/Request/AddProviderRequest.cs
@@ -1,14 +1,15 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace BalanceApi.Model.ViewModels
+namespace BalanceApi.Model.Views.Request
 {
-    public class AddTransactionTypeRequest
+    public class AddProvider
     {
         [Required]
         [StringLength(100)]
         public string Name { get; set; }
         
         [Required]
-        public bool Credit { get; set; }
+        [StringLength(2)]
+        public string Country { get; set; }
     }
 }
